changes in weather component
