+++ src/components/weather/Weather.tsx
@@ -1,12 +1,33 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 import { Style } from "./Weather.style";
 import { API_ICON_ULR } from "../../API/API";
 import { useParams } from "react-router-dom";
+import useGeoLocation from "../../useGeoLocation";
+import { API_URL, API_KEY } from "../../API/API";
+import Footer from "../footer/Footer";
 
-const Weather: React.FunctionComponent<any> = ({ data, day }) => {
-  const { city, date } = useParams();
-  const currentCityName = data?.city?.name;
+const t = new Date();
+const dd = String(t.getDate()).padStart(2, "0");
+const mm = String(t.getMonth() + 1).padStart(2, "0");
+const yyyy = t.getFullYear();
+const today = String(yyyy + "-" + mm + "-" + dd);
 
+const Weather: React.FunctionComponent<any> = () => {
+  const { date } = useParams();
+  const location = useGeoLocation();
+  const [data, setData] = useState<any>();
+  const lat = JSON.stringify(location?.coordinates?.lat);
+  const lon = JSON.stringify(location?.coordinates?.lon);
+
+  useEffect(() => {
+    if (lat && lon) {
+      fetch(`${API_URL}forecast?lat=${lat}&lon=${lon}&appid=${API_KEY}`)
+        .then((response) => response.json())
+        .then((results) => {
+          setData(results);
+        });
+    }
+  }, [lat, lon]);
   const result = data?.list?.filter((item: any) => {
     return item.dt_txt.includes(date);
   });
@@ -14,11 +35,8 @@ const Weather: React.FunctionComponent<any> = ({ data, day }) => {
 
   return (
     <Style.Content>
-      <h1>{city}</h1>
-      <h1>{date}</h1>
-
       <Style.CurrentCard>
-        <h3>{currentCityName}</h3>
+        <h3>{data?.city?.name}</h3>
         <h1>
           {Math.ceil(((data?.list?.[0]?.main?.temp - 273) * 9) / 5 + 32) + "°F"}
         </h1>
@@ -31,7 +49,7 @@ const Weather: React.FunctionComponent<any> = ({ data, day }) => {
       <Style.WeatherHour>
         {data?.list
           ?.filter((item: any) => {
-            return item.dt_txt.includes(day);
+            return item.dt_txt.includes(today);
           })
           .map((item: any) => {
             return (
@@ -50,6 +68,7 @@ const Weather: React.FunctionComponent<any> = ({ data, day }) => {
             );
           })}
       </Style.WeatherHour>
+      <Footer data={data} />
     </Style.Content>
   );
 };
